Refactor common user review spec helper methods
Several spec helper methods for user reviews were duplicated across
files. This commit extracts those methods to a spec helper module.
Additionally, this commit refactors some other specs to use these helper
methods, instead of more difficult-to-read CSS assertions.
