+++ spec/features/admin_manages_user_review_spec.rb
@@ -47,22 +47,4 @@ feature "Admin manages a user review for an event" do
   def click_update_button
     click_button "Update Review"
   end
-
-  def expect_user_review_score_to_update_from(old_value, new_value)
-    expect(page).not_to have_user_review_score(old_value)
-    expect(page).to have_user_review_score(new_value)
-  end
-
-  def expect_user_review_body_to_update_from(old_value, new_value)
-    expect(page).not_to have_user_review_body(old_value)
-    expect(page).to have_user_review_body(new_value)
-  end
-
-  def have_user_review_score(score)
-    have_css(".user-review .review-score", text: score)
-  end
-
-  def have_user_review_body(text)
-    have_css(".user-review .review-content", text: text)
-  end
 end
