+++ spec/features/user_edits_user_review_spec.rb
@@ -34,22 +34,4 @@ feature "User can edit a user review from an event page" do
   def visit_event
     visit event_path(@event, as: @user)
   end
-
-  def expect_user_review_score_to_update_from(old_value, new_value)
-    expect(page).not_to have_user_review_score(old_value)
-    expect(page).to have_user_review_score(new_value)
-  end
-
-  def expect_user_review_body_to_update_from(old_value, new_value)
-    expect(page).not_to have_user_review_body(old_value)
-    expect(page).to have_user_review_body(new_value)
-  end
-
-  def have_user_review_score(score)
-    have_css(".user-review .review-score", text: score)
-  end
-
-  def have_user_review_body(text)
-    have_css(".user-review .review-content", text: text)
-  end
 end
